+++ server/api/api/settings.py
@@ -44,6 +44,7 @@ INSTALLED_APPS = [
     'rest_framework_swagger',
     # internal apps
     'user',
+    'mathcaptcha',
     'registration',
     # 'thread',
