+++ server/api/mathcaptcha/urls.py
@@ -1,0 +1,8 @@
+from django.conf.urls import url
+
+from .views import GetCaptchaView
+
+
+urlpatterns = [
+    url(r'^$', GetCaptchaView.as_view(), name='get-captcha'),
+]
