+++ server/api/registration/urls.py
@@ -1,9 +1,8 @@
 from django.conf.urls import url
 
-from .views import RegistrationView, GetCaptchaView
+from .views import RegistrationView
 
 
 urlpatterns = [
     url(r'^$', RegistrationView.as_view(), name='registration'),
-    url(r'^captcha/$', GetCaptchaView.as_view(), name='get-captcha'),
 ]
